refactor: reuse native app for the wasm bindings
