+++ Cargo.toml
@@ -217,7 +217,7 @@ tracing-subscriber = "0.3.19"
 tracing-test = { version = "0.2.5", features = ["no-env-filter"] }
 trait-variant = "0.1"
 twox-hash = "2.1.1"
-uniffi = "0.31.1"
+uniffi = "0.31.0"
 url = "2.5.0"
 uuid = { version = "1.15.1", features = ["v4"] }
 wasm-bindgen-futures = "0.4"
