+++ crates/service/src/lib.rs
@@ -33,9 +33,9 @@
 //! # })
 //! ```
 
-#![cfg(not(target_arch = "wasm32"))]
-
 pub mod errors;
+
+#[cfg(not(target_arch = "wasm32"))]
 pub mod service;
 
 #[cfg(feature = "session")]
@@ -48,4 +48,5 @@ pub use slim_datapath::messages::utils::SlimHeaderFlags;
 pub use errors::ServiceError;
 #[cfg(feature = "session")]
 pub use errors::SubscriptionAckError;
+#[cfg(not(target_arch = "wasm32"))]
 pub use service::{KIND, Service, ServiceBuilder, ServiceConfiguration};
